Move remaining channel handlers into separate files and delete NIO.swift
The examples are no longer needed because we have end to end tests which show how to setup a client and server
